[악어] 2,3 단계 자동 DI 미션 제출합니다 (#114)
수고하셨습니다 악어~ 🐊 이번 2,3단계 미션을 진행하면서 DI뿐만 아니라 뷰모델의 생명주기에 대해서 같이 논의하고 많이 얻어간 것 같아서 보람차네요 ㅎㅎ

제 피드백을 반영하시면서 리팩토링과 수정할 부분이 더 많이 생긴 것 같지만.. 4단계 미션을 위해 일단 머지하겠습니다. 먼저 4단계에 집중해 주시고, 새로 드린 피드백들은 4단계 이후에 반영하시면 좋을 것 같아요!

저번 코멘트에 대한 답글도 확인 부탁드릴게요~

+++ app/src/main/java/woowacourse/shopping/data/CartProductDao.kt
@@ -13,5 +13,5 @@ interface CartProductDao {
     suspend fun insert(cartProduct: CartProductEntity)
 
     @Query("DELETE FROM cart_products WHERE id = :id")
-    suspend fun delete(id: Long)
+    suspend fun delete(id: Int)
 }
